Fix bpmn palette missing provider error in prod mode
